fix saving

+++ engine/inc/C4Object.h
@@ -60,30 +60,30 @@ public:
 public:
   uint32_t Magic;
   C4ID id;
-  long Number;
-  long Status;       // NoSave //
-  long RemovalDelay; // NoSave //
-  long Owner;
-  long Controller;
-  long Category;
-  long x, y, r;
-  long motion_x, motion_y;
-  long NoCollectDelay;
-  long MaskPutMotion;
-  long Base;
-  long Con;
-  long Mass;
-  long Damage;
-  long Energy;
-  long MagicEnergy;
-  long Alive;
-  long Breath;
-  long FirePhase;
-  long InMat; // SyncClearance-NoSave //
-  long Color;
-  long Timer;
-  long ViewEnergy;           // NoSave //
-  long Audible, LastAudible; // NoSave //
+  int Number;
+  int Status;       // NoSave //
+  int RemovalDelay; // NoSave //
+  int Owner;
+  int Controller;
+  int Category;
+  int x, y, r;
+  int motion_x, motion_y;
+  int NoCollectDelay;
+  int MaskPutMotion;
+  int Base;
+  int Con;
+  int Mass;
+  int Damage;
+  int Energy;
+  int MagicEnergy;
+  int Alive;
+  int Breath;
+  int FirePhase;
+  int InMat; // SyncClearance-NoSave //
+  int Color;
+  int Timer;
+  int ViewEnergy;           // NoSave //
+  int Audible, LastAudible; // NoSave //
   long Local[C4MaxVariable];
   FIXED fix_x, fix_y, fix_r; // SyncClearance-Fix //
   FIXED xdir, ydir, rdir;
